Changes to be committed: 	deleted:    src/App.js 	renamed:    src/Card.js -> src/components/Card.js 	renamed:    src/CardList.js -> src/components/CardList.js 	new file:   src/components/Scroll.js 	renamed:    src/SearchBox.js -> src/components/SearchBox.js 	renamed:    src/App.css -> src/containers/App.css 	new file:   src/containers/App.js 	renamed:    src/SEGA.woff -> src/containers/SEGA.woff 	modified:   src/index.js 	deleted:    src/robots.js
