+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
+import App from './containers/App';
 import 'tachyons';
 
 ReactDOM.render(<App />,document.getElementById('root'));
